freelancer-backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.2
python-jose==3.3.0
passlib==1.7.4
pydantic==1.10.7